Add UITextField and UITextView Refactoring code
